monitor

+++ app/api/monitor/[id]/route.ts
@@ -1,8 +1,15 @@
+import { checkAdminRules } from '@/lib/check-admin-rules';
 import { prisma } from '@/prisma/prisma-client';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
   try {
+    const adminRules = await checkAdminRules(true);
+
+    if (adminRules.status !== 200) {
+      return NextResponse.json({ message: adminRules.message }, { status: adminRules.status });
+    }
+
     await prisma.monitor.delete({
       where: {
         id: Number(params.id),
